removing one footer.js file and combining with other index.js file
